ml-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
tensorflow==2.15.0
xgboost==2.0.2
mlflow==2.8.1
redis==5.0.1
psycopg2-binary==2.9.9
python-multipart==0.0.6
pydantic==2.5.0
httpx==0.25.2
python-dotenv==1.0.0
prometheus-client==0.19.0
schedule==1.2.0
websockets==12.0
requests==2.31.0